Add proper react error handler
